[#626] TextView Extension 분리
